Rename mock file
