+++ ynca/__init__.py
@@ -2,7 +2,7 @@ import logging
 from typing import Dict, List
 
 
-from .connection import YncaConnection, ynca_console
+from .connection import YncaConnection, YncaConnectionError, ynca_console
 from .constants import DSP_SOUND_PROGRAMS, Mute, Subunit, ZONES
 from .receiver import Receiver
